Complete the Attached Properties for password box
Complete the Attached Properties for password box.
1st we learn how attached property is done in Standard way.
Then we learn how to use the generics to first create a BaseAttachedProperty class. Which is our base class for attached properties.
Then we use our base class to create HasText property and MonitorPassword property.
After doing these things, we hook up the visibility password box with these 2 propertes.
But we run into an issue.
Which is solved by using converter method.
In this converter class we just use a boolean.